init subtask3 completed

+++ src/main/kotlin/subtask2/Combinator.kt
@@ -16,7 +16,7 @@ class Combinator {
             }
         }
 
-        return if (resultVal == n) null else resultVal.toInt()
+        return if (resultVal == n) null else resultVal
     }
 
     fun factorial(n: Int): Double = if (n < 2) 1.0 else n * factorial(n - 1)
